64 bit fixes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@55534 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/osx/cocoa/window.mm
@@ -324,7 +324,7 @@ void SetupMouseEvent( wxMouseEvent &wxevent , NSEvent * nsEvent )
         
         wxRegion updateRgn;
         const NSRect *rects;
-        int count ;
+        NSInteger count;
 
         [self getRectsBeingDrawn:&rects count:&count];
         for ( int i = 0 ; i < count ; ++i )
